diffusers/cache_diffusion/requirements.txt
accelerate>=0.22.0
cuda-python
datasets>=2.14.5
diffusers>=0.27.2
opencv-python==4.8.1.78
optimum
peft>=0.10.0
polygraphy==0.49.9
transformers>=4.40.2
